+++ src/main/java/com/example/lab1/entity/IBehaviour.java
@@ -1,11 +1,11 @@
 package com.example.lab1.entity;
 
 public interface IBehaviour {
-    int getX();
-    int getY();
-    int getSpeed();
-    void setX(int x);
-    void setY(int y);
+    double getX();
+    double getY();
+    double getSpeed();
+    void setX(double x);
+    void setY(double y);
     Direction getDirection();
     void setDirection(Direction direction);
     void updateDirection();
